+++ my-travel-map/src/utils/fetchApi.js
@@ -1,0 +1,6 @@
+// Get All APIs
+export const fetchApi = async (url) => {
+    const res = await fetch(url, { cache: "no-store" });
+    if (!res.ok) throw new Error("Failed to fetch");
+    return res.json();
+};
